Alerts UI standalone page
Summary:
Created with create-react-app. This provides for a standalone UI to be shipped with the prometheus-configmanager repo. The docker-compose file will spin up the entire service with default configurations, making it easy to test and try out.

I will update the helm chart in another diff to add the UI container to it.

A few things need to be added in the near term:
* thresholdEditor
* Tenant selector

And a few more are not needed immediately, but will be added later:
* Routes
* Suppressions

allow-large-files

Reviewed By: aclave1

Differential Revision: D21867726

fbshipit-source-id: 4ab3c9bb85bf9cce152c8f583731fb207fb36247