Add PagerDuty Support to AlertManager (#1767)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1767

This diff provides the functionality to use PagerDuty with AlertManager, in addition to unit testing and documentation for the new feature.

Note: this did not have any strange fields like the duration field for the PushoverConfig. I used existing fields for the images and links fields (in the prometheus config), as all the underlying structures were string, and no modification was required.

Reviewed By: Scott8440

Differential Revision: D21871913

fbshipit-source-id: d679f3c8d50b49f3777e310a7e007144e42fdb71