Individual Template editing
Summary:
Adds functionality to edit individual template strings in template files.

Utilizes the `text/template` library to handle parsing which wouldn't be possible (without a lot of work) by a frontend client.

With this you can get a list of all template strings that are availble, and add/edit/remove them individually, which makes creating a UI around this very simple.

Reviewed By: xjtian

Differential Revision: D22174867

fbshipit-source-id: 52447194c6b5e1f138cd30779a4fe68949693ca9