Support week and day in rule durations (#24)
Summary:
Pull Request resolved: https://github.com/facebookincubator/prometheus-configmanager/pull/24

Reads were failing when a rule duration was manually configured with "w" and "d" (weeks and days). This should theoretically work even though there isn't an option in the UI to do that yet. The cause was converting from prometheus' `model.Duration` to the built-in `time.Duration` when building the JSON to return in the request. Turns out `model.Duration` also has a string() function we can use, so just use that instead. buit-in time doesn't handle w/d which was the problem.

Reviewed By: aclave1

Differential Revision: D25450686

fbshipit-source-id: 794a945082b1b0c6e151d95560086de1836d5c93